# backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API de Reserva de Espacios",
        "description": "API REST para gestión de espacios y reservas",
        "contact": {
            "email": "soporte@reservaespacios.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api",
            "description": "Servidor de desarrollo"
        }
    ],
    "paths": {
...
        "/espacios": {
            "get": {
                "tags": [
                    "Espacios"
                ],
                "summary": "Listar todos los espacios",
                "operationId": "8eda707d36e5fd9525b329abf637161c",
                "parameters": [
                    {
                        "name": "tipo",
                        "in": "query",
                        "description": "Filtrar por tipo de espacio",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "disponible",
                        "in": "query",
                        "description": "Filtrar por disponibilidad (0 o 1)",
                        "required": false,
                        "schema": {
                            "type": "boolean"
                        }
                    },
                    {
                        "name": "capacidad_minima",
                        "in": "query",
                        "description": "Filtrar por capacidad mínima",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Lista de espacios",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "id": {
                                                "type": "integer",
                                                "example": 1
                                            },
                                            "nombre": {
                                                "type": "string",
                                                "example": "Sala de Conferencias A"
                                            },
                                            "descripcion": {
                                                "type": "string"
                                            },
                                            "capacidad": {
                                                "type": "integer",
                                                "example": 50
                                            },
                                            "tipo": {
                                                "type": "string",
                                                "example": "Sala de Conferencias"
                                            },
                                            "imagen_url": {
                                                "type": "string"
                                            },
                                            "disponible": {
                                                "type": "boolean"
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Espacios"
                ],
                "summary": "Crear nuevo espacio",
                "operationId": "9d4728bea37e8c494ddaa6ae5bcd924a",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "nombre",
                                    "capacidad",
                                    "tipo"
                                ],
                                "properties": {
                                    "nombre": {
                                        "type": "string",
                                        "example": "Sala de Conferencias A"
                                    },
                                    "descripcion": {
                                        "type": "string",
                                        "example": "Sala amplia con proyector"
                                    },
                                    "capacidad": {
                                        "type": "integer",
                                        "example": 50
                                    },
                                    "tipo": {
                                        "type": "string",
                                        "example": "Sala de Conferencias"
                                    },
                                    "imagen_url": {
                                        "type": "string",
                                        "example": "https://example.com/imagen.jpg"
                                    },
                                    "disponible": {
                                        "type": "boolean",
                                        "example": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Espacio creado exitosamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "mensaje": {
                                            "type": "string"
                                        },
                                        "espacio": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "No autenticado"
                    },
                    "422": {
                        "description": "Errores de validación"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/espacios/{id}": {
            "get": {
                "tags": [
                    "Espacios"
                ],
                "summary": "Obtener un espacio específico",
                "operationId": "fbdf656ad17337ad920bb06a263ab169",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del espacio",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Detalles del espacio",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer"
                                        },
                                        "nombre": {
                                            "type": "string"
                                        },
                                        "descripcion": {
                                            "type": "string"
                                        },
                                        "capacidad": {
                                            "type": "integer"
                                        },
                                        "tipo": {
                                            "type": "string"
                                        },
                                        "imagen_url": {
                                            "type": "string"
                                        },
                                        "disponible": {
                                            "type": "boolean"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Espacio no encontrado"
                    }
                }
            },
            "put": {
                "tags": [
                    "Espacios"
                ],
                "summary": "Actualizar un espacio",
                "operationId": "b7fdfc1411fc7ce693dcbd01eb7c61cb",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del espacio",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "nombre": {
                                        "type": "string"
                                    },
                                    "descripcion": {
                                        "type": "string"
                                    },
                                    "capacidad": {
                                        "type": "integer"
                                    },
                                    "tipo": {
                                        "type": "string"
                                    },
                                    "imagen_url": {
                                        "type": "string"
                                    },
                                    "disponible": {
                                        "type": "boolean"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Espacio actualizado exitosamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "mensaje": {
                                            "type": "string"
                                        },
                                        "espacio": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "No autenticado"
                    },
                    "404": {
                        "description": "Espacio no encontrado"
                    },
                    "422": {
                        "description": "Errores de validación"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Espacios"
                ],
                "summary": "Eliminar un espacio",
                "operationId": "301c3c1cad75967459c3858bd14f3909",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del espacio",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Espacio eliminado exitosamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "mensaje": {
                                            "type": "string",
                                            "example": "Espacio eliminado exitosamente"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "No autenticado"
                    },
                    "404": {
                        "description": "Espacio no encontrado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}